+++ src/automation/azext_automation/aaz/latest/automation/runtime_environment/_create.py
@@ -22,9 +22,9 @@ class Create(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2023-05-15-preview",
+        "version": "2024-10-23",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.automation/automationaccounts/{}/runtimeenvironments/{}", "2023-05-15-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.automation/automationaccounts/{}/runtimeenvironments/{}", "2024-10-23"],
         ]
     }
 
@@ -182,7 +182,7 @@ class Create(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2023-05-15-preview",
+                    "api-version", "2024-10-23",
                     required=True,
                 ),
             }